requirements.txt
Django
django-paypal
django-bootstrap3
Pillow